expand and improve USER_LOGIN action in reducer
Users now have IDs to improve finding the logged in user in the users
array in the store.

Previously, if the username the user entered existed in state, and the
password they entered matched any user's password, they could
log in. Regardless if the password belonged to the entered username.
This is now fixed.